ai-notebook-backend/requirements.txt
fastapi
uvicorn
python-multipart
openai
PyMuPDF
faiss-cpu
tiktoken
python-dotenv
pydantic
numpy
requests
streamlit